fix(cumulant_discovery): re-narrow before preserve_nonzero float32 underflow guard
The preserve_nonzero guard in _require_float32 tested the un-narrowed
stored value, so for built-in float inputs (the normal caller path for
gamma/replacement_rate/predictor_step_size) a value that underflows only
in float32 (e.g. 1e-50) was silently accepted instead of rejected.
Re-narrow via float(np.float32(stored)) at the guard, mirroring
option_value_duration.py, so the check fires identically for built-in
float and np.float64. Returned value is unchanged.

Closes #2720

+++ alberta_framework/core/cumulant_discovery.py
@@ -106,7 +106,7 @@ def _require_float32(
         upper_inclusive=upper_inclusive,
         positive=positive,
     )
-    if preserve_nonzero and numerator != 0 and stored == 0.0:
+    if preserve_nonzero and numerator != 0 and float(np.float32(stored)) == 0.0:
         raise ValueError(f"{name} must remain nonzero once narrowed to float32")
     return stored
